Fixed bug in sac struct and functions changing long with gint32 and ulong with guint32
git-svn-id: svn+ssh://svn.rm.ingv.it/svn/teseo2/trunk@781 afdc3c29-bc1c-424d-8b75-1d331506fbbc

+++ teseo-2/src/teseocallbacks.c
@@ -626,8 +626,8 @@ on_sac1_activate                       (GtkMenuItem     *menuitem,
     float paper_velocity = 1;
     gchar * filename = NULL;
     gint result;
-    long NZYEAR=1968, NZJDAY=338, NZHOUR=12, NZMIN=0, NZSEC=0, NZMSEC=0;
-    long MONTH=0, DAY=0;
+    gint32 NZYEAR=1968, NZJDAY=338, NZHOUR=12, NZMIN=0, NZSEC=0, NZMSEC=0;
+    gint32 MONTH=0, DAY=0;
     GDate* jdate;
 
     // TODO catch scale value
